+++ .tools/test-geosphere-at.php
@@ -120,6 +120,19 @@ class IPSModule
     public function Create()
     {
     }
+    public function MaintainVariable($ident, $name, $type, $profile, $position, $keep = true)
+    {
+    }
+    public function SetValue($ident, $value)
+    {
+    }
+}
+
+const VARIABLETYPE_STRING = 3;
+const VARIABLETYPE_INTEGER = 1;
+function IPS_VariableProfileExists(string $name): bool
+{
+    return true;
 }
 
 require __DIR__ . '/../WarnHub/module.php';